fix(core): move separate observablelogic to useDocumentVersions
